# apps/api/agent.py
"""Onboarding-agent endpoints — drive the agent (a platform plugin) from React.

Uses the agent's headless AgentSession (same conversation code as the Streamlit
plugin) with a per-session store. Upload a working folder (.zip) to propose a
mapping, then send messages to walk the decisions, build, and ask the graph.
"""
from __future__ import annotations

import os
import shutil
import sys
import tempfile
import uuid
import zipfile
from collections import OrderedDict
from pathlib import Path
from typing import Any

from fastapi import APIRouter, Depends, File, Form, HTTPException, UploadFile
from pydantic import BaseModel

from backend.workspace import WorkspaceContext

from .deps import get_ctx

# The agent lives as a mounted module; put it on the path so we can import it.
_MODULES = os.getenv("MODULES_DIR", "/app/data/modules")
if _MODULES not in sys.path:
    sys.path.insert(0, _MODULES)

router = APIRouter(prefix="/api/workspaces/{workspace_id}/agent", tags=["agent"])

# session_id -> AgentSession. In-memory, one process, and *bounded*: an LRU
# capped at $AGENT_SESSION_CAP (default 32). Every session the store holds
# keeps a live LLM conversation plus (after an upload) a temp working folder,
# so an unbounded dict leaks both under any real traffic.
_SESSIONS: "OrderedDict[str, Any]" = OrderedDict()


def _session_cap() -> int:
    try:
        return max(1, int(os.getenv("AGENT_SESSION_CAP", "32")))
    except ValueError:
        return 32


def _dispose(sess: Any) -> None:
    """Release an evicted session's on-disk footprint (its upload tmp dir)."""
    tmp = getattr(sess, "_upload_tmp", None)
    if tmp:
        shutil.rmtree(tmp, ignore_errors=True)


def _put(session_id: str, sess: Any) -> None:
    _SESSIONS[session_id] = sess
    _SESSIONS.move_to_end(session_id)
    while len(_SESSIONS) > _session_cap():
        _, oldest = _SESSIONS.popitem(last=False)
        _dispose(oldest)


def _new_session(ctx: WorkspaceContext):
    from onboarding_agent.headless import AgentSession
    from .deps import ws_root
    return AgentSession(
        ctx.id, str(ws_root(ctx)), ctx, ctx.graphdb_repository or ctx.id,
        model=os.getenv("LLM_MODEL", "sonnet"),
    )


def _get(session_id: str, ctx: WorkspaceContext):
    """The session for ``session_id`` — but ONLY if it belongs to ``ctx``.

    ``get_ctx`` on the route only proves the caller may see the workspace NAMED IN THE
    URL; ``_SESSIONS`` is a single process-wide dict keyed purely by session_id, so
    without this check a caller who knows (or is handed) a session_id created for a
    DIFFERENT workspace could drive that session — including one on a private workspace
    they have no visibility into — merely by calling through a URL for a workspace they
    DO have access to. 404 (not 403) either way, so a mismatch can't be used to probe
    whether a given session_id exists."""
    sess = _SESSIONS.get(session_id)
    if sess is None or sess.ws_id != ctx.id:
        raise HTTPException(status_code=404, detail="agent session not found — start a new one")
    _SESSIONS.move_to_end(session_id)  # touched → most recently used
    return sess


class StartBody(BaseModel):
    model: str | None = None
    chat_id: str | None = None


@router.get("/models")
def models() -> list[dict[str, str]]:
    from onboarding_agent.headless import MODELS
    return MODELS


@router.get("/chats")
def chats(ctx: WorkspaceContext = Depends(get_ctx)) -> list[dict[str, Any]]:
    """Persisted conversations for this workspace."""
    return _new_session(ctx).list_chats()


@router.post("/session")
def start_session(body: StartBody | None = None, ctx: WorkspaceContext = Depends(get_ctx)) -> dict[str, Any]:
    """Open a conversation — fresh, or load a persisted chat_id; optional model."""
    sid = uuid.uuid4().hex
    sess = _new_session(ctx)
    if body and body.model:
        sess.set_model(body.model)
    _put(sid, sess)
    if body and body.chat_id:
        sess.load_chat(body.chat_id)
    return {"session_id": sid, **sess.snapshot()}


class ModelBody(BaseModel):
    session_id: str
    model: str


@router.post("/model")
def set_model(body: ModelBody, ctx: WorkspaceContext = Depends(get_ctx)) -> dict[str, str]:
    _get(body.session_id, ctx).set_model(body.model)
    return {"model": body.model}


class ModeBody(BaseModel):
    session_id: str
    mode: str                # "auto" (apply defaults) | "manual" (walk decisions)


@router.post("/mode")
def set_mode(body: ModeBody, ctx: WorkspaceContext = Depends(get_ctx)) -> dict[str, str]:
    sess = _get(body.session_id, ctx)
    sess.set_mode(body.mode)
    return {"mode": sess.state.oa_mode}


class ApiKeyBody(BaseModel):
    provider: str            # "anthropic" | "mistral"
    key: str | None = None   # falsy clears the stored key


@router.post("/api-key")
def set_api_key(body: ApiKeyBody, ctx: WorkspaceContext = Depends(get_ctx)) -> dict[str, Any]:
    """Set (and persist) an LLM provider key from the settings panel; returns the
    updated settings (key status never echoes the key itself)."""
    if body.provider not in ("anthropic", "mistral"):
        raise HTTPException(status_code=400, detail="provider must be 'anthropic' or 'mistral'")
    sess = _new_session(ctx)
    sess.set_api_key(body.provider, (body.key or "").strip() or None)
    return sess.settings()


@router.get("/settings")
def settings(ctx: WorkspaceContext = Depends(get_ctx)) -> dict[str, Any]:
    """Settings-panel data: per-provider key status + the read-only folder where the
    agent stores this workspace's artifacts (chats, ingestion report, provenance)."""
    return _new_session(ctx).settings()


class Message(BaseModel):
    session_id: str
    text: str


@router.post("/message")
def message(body: Message, ctx: WorkspaceContext = Depends(get_ctx)) -> dict[str, Any]:
    return _get(body.session_id, ctx).send(body.text)


def _stream_response(sess, text: str, ctx: WorkspaceContext):
    return _sse(lambda: sess.send_stream(text), ctx)


def _sse(produce, ctx: WorkspaceContext):
    """Server-sent events for one agent step: ``produce()`` yields ``(kind, data)``
    pairs (``token``, ``result``); a ``done`` event ends the stream."""
    import json as _json
    import logging
    import queue as _queue
    import threading
    from fastapi.responses import StreamingResponse

    # The turn runs in its own thread feeding a queue, so the response
    # generator can emit `: keepalive` comments while the agent is silent.
    # Long deterministic phases (harvest, build, OM runs) send no tokens for
    # minutes — an idle SSE stream gets killed by middleboxes between the
    # browser and the api (seen live: EventSource errored mid-`build` while
    # the build finished server-side). Comment lines are ignored by
    # EventSource. Side benefit: the turn now survives a client disconnect —
    # the worker runs to completion instead of dying with the response.
    q: "_queue.Queue" = _queue.Queue()

    def worker():
        try:
            for kind, data in produce():
                q.put((kind, data))
        except Exception:
            logging.getLogger("digicities.agent").exception("agent turn failed")
        finally:
            # An agent turn can build/import/reset the workspace — and it rides
            # a GET that outlives the push middleware, so publish here, after
            # the turn actually finished (no-op for local-fs workspaces).
            from backend.workspace import mirror
            mirror.push(ctx)
            q.put(None)

    threading.Thread(target=worker, daemon=True).start()

    def gen():
        while True:
            try:
                item = q.get(timeout=15)
            except _queue.Empty:
                yield ": keepalive\n\n"
                continue
            if item is None:
                yield "event: done\ndata: {}\n\n"
                return
            kind, data = item
            yield f"event: {kind}\ndata: {_json.dumps(data)}\n\n"

    return StreamingResponse(gen(), media_type="text/event-stream",
                             headers={"Cache-Control": "no-cache", "X-Accel-Buffering": "no"})


@router.get("/message/stream")
def message_stream(session_id: str, text: str, ctx: WorkspaceContext = Depends(get_ctx)):
    """Server-sent events: `token` as the LLM writes, then `result` with the full turn.

    GET exists for EventSource clients; long messages should use the POST
    variant so the text rides in the body, not the query string / access logs.
    """
    return _stream_response(_get(session_id, ctx), text, ctx)


@router.post("/message/stream")
def message_stream_post(body: Message, ctx: WorkspaceContext = Depends(get_ctx)):
    """Same SSE stream, message in the request body (fetch + ReadableStream)."""
    return _stream_response(_get(body.session_id, ctx), body.text, ctx)


def _propose_or_defer(sess, folder, propose: bool) -> dict[str, Any]:
    """Map the uploaded folder now, or (``propose=false``) only record it and
    return at once: the client then runs the mapping over ``/propose/stream``.
    Mapping takes minutes and sends nothing meanwhile, and networks cut
    requests that stay silent that long (seen live at 60 s): the upload
    errored in the browser while the server carried on. The stream sends a
    keepalive every 15 s instead."""
    if propose:
        return sess.propose(Path(folder))
    sess._pending_propose = str(folder)
    return {**sess.snapshot(), "pending_propose": True}


@router.get("/propose/stream")
def propose_stream(session_id: str, ctx: WorkspaceContext = Depends(get_ctx)):
    """Server-sent events: map the folder recorded by ``upload(propose=false)``,
    then ``result`` with the full state and ``done``. 409 when there is nothing
    waiting to be mapped."""
    sess = _get(session_id, ctx)
    folder = getattr(sess, "_pending_propose", None)
    if not folder:
        raise HTTPException(status_code=409, detail="no uploaded folder is waiting to be mapped")
    sess._pending_propose = None

    def produce():
        yield "result", sess.propose(Path(folder))

    return _sse(produce, ctx)


@router.get("/state")
def state(session_id: str, ctx: WorkspaceContext = Depends(get_ctx)) -> dict[str, Any]:
    return _get(session_id, ctx).snapshot()


@router.get("/commands")
def commands(session_id: str, ctx: WorkspaceContext = Depends(get_ctx)) -> dict[str, Any]:
    """The commands that can run in the conversation's current step, for the
    chat's command list: ``{"stage", "commands": [{key, area, form, usage,
    description, when_to_use, example, example_result, states, slots: [{name,
    kind, label, help, choices: [{value, label}]}]}]}``. Each command says when to
    use it and what its example does; each slot says what goes in it. Slot
    choices come from the workspace (classes, attributes, scenarios, instances,
    the numbered decisions) and the ontology (the link predicates under
    ``linksComponent``). Generated from the agent's command registry, the same
    list its router and help use."""
    return _get(session_id, ctx).commands()


def _continuing(sess) -> bool:
    """Is this chat in the middle of an onboarding? Its own upload exists (even
    if a restart wiped the temp dir), or it carries a mapping (a chat restored
    after a restart). Only then does a new upload ADD to the working folder. A
    fresh chat on a workspace that was onboarded before starts a fresh folder:
    nested into the previous onboarding's kept copy, the agent's "replace" would
    onboard old and new together (and wipe the new upload with the old copy)."""
    if getattr(sess, "_upload_folder", None):
        return True
    st = getattr(sess, "state", None)
    return bool(getattr(st, "oa_spec", None)) \
        or getattr(st, "oa_stage", "start") not in ("start", None)


def _working_folder(sess) -> str | None:
    """The folder this chat reads, if it still exists. The upload's temp dir is
    wiped by a server restart; the agent keeps a copy of the folder inside the
    workspace and names it (``working_folder``, absent on older agents), so an
    upload that ADDS a file after a restart lands next to the data it belongs
    with instead of starting a one-file folder."""
    if not _continuing(sess):
        return None
    existing = getattr(sess, "_upload_folder", None)
    if existing and Path(existing).is_dir():
        return existing
    finder = getattr(sess, "working_folder", None)
    if callable(finder):
        try:
            got = finder()
        except Exception:
            got = None
        if got and Path(got).is_dir():
            return str(got)
    return None


@router.post("/upload")
def upload(
    session_id: str = Form(...),
    file: UploadFile = File(...),
    propose: bool = Form(True),
    ctx: WorkspaceContext = Depends(get_ctx),
) -> dict[str, Any]:
    """Upload a working folder as a .zip, OR a single file. When a working folder already exists
    in the session, more data ACCUMULATES into it: a second .zip is nested under a subfolder, a
    single file is dropped in (e.g. an onboarding guide, or a file a previous read missed) — then
    the folder is re-read. With no prior folder, the upload becomes the working folder (a .zip's
    contents, or a one-file folder). Start fresh = New chat. The agent proposes a mapping.

    A plain ``def`` on purpose, like every other agent endpoint: FastAPI runs it
    in a worker thread. The proposal takes minutes; as ``async def`` it ran on
    the event loop and froze EVERY request meanwhile — the health probe timed
    out, Kubernetes restarted the container, and every chat session was lost."""
    sess = _get(session_id, ctx)
    if not file.filename:
        raise HTTPException(status_code=400, detail="No file was uploaded")
    name = Path(file.filename).name            # basename only — no path traversal
    data = file.file.read()

    # ── a .zip → extract (with a zip-slip guard) ────────────────────────────────
    if name.lower().endswith(".zip"):
        scratch = Path(tempfile.mkdtemp(prefix="oa-zip-"))
        zpath = scratch / "upload.zip"
        zpath.write_bytes(data)
        root = (scratch / "x").resolve()
        try:
            with zipfile.ZipFile(zpath) as z:
                for m in z.infolist():
                    # zip-slip guard: no entry may land outside the extract root
                    target = (root / m.filename).resolve()
                    if root != target and root not in target.parents:
                        raise HTTPException(
                            status_code=400,
                            detail=f"Zip entry escapes its root: {m.filename!r}")
                z.extractall(root)
        except HTTPException:
            shutil.rmtree(scratch, ignore_errors=True)
            raise
        except Exception as exc:
            shutil.rmtree(scratch, ignore_errors=True)
            raise HTTPException(status_code=400, detail=f"Bad zip: {exc}") from exc
        # If the zip had a single top-level folder, descend into it.
        entries = [p for p in root.iterdir() if not p.name.startswith("__MACOSX")]
        content = entries[0] if len(entries) == 1 and entries[0].is_dir() else root

        existing = _working_folder(sess)
        if existing:
            # ADD MORE DATA: nest this zip under a subfolder of the current working folder so
            # its files accumulate (the intake reads the folder recursively). Start fresh = New chat.
            base = Path(name).stem or "added"
            dest = Path(existing) / base
            i = 2
            while dest.exists():
                dest = Path(existing) / f"{base}-{i}"
                i += 1
            shutil.copytree(content, dest)
            shutil.rmtree(scratch, ignore_errors=True)
            sess.state.oa_messages.append(("user", f"📦 Added `{name}` to the working folder"))
            return _propose_or_defer(sess, Path(existing), propose)

        # a fresh working folder (no prior upload in this session)
        prev = getattr(sess, "_upload_tmp", None)
        if prev:
            shutil.rmtree(prev, ignore_errors=True)
        sess._upload_tmp = str(scratch)
        sess._upload_folder = str(content)
        sess.state.oa_messages.append(("user", f"📦 Uploaded `{name}`"))
        return _propose_or_defer(sess, content, propose)

    # ── a single file added to the current working folder → re-read it ───────────
    existing = _working_folder(sess)
    if existing:
        (Path(existing) / name).write_bytes(data)
        sess.state.oa_messages.append(("user", f"📎 Added `{name}` to the working folder"))
        return _propose_or_defer(sess, Path(existing), propose)

    # ── a single file with no prior folder → a one-file working folder ───────────
    prev = getattr(sess, "_upload_tmp", None)
    if prev:
        shutil.rmtree(prev, ignore_errors=True)
    tmp = Path(tempfile.mkdtemp(prefix="oa-upload-"))
    sess._upload_tmp = str(tmp)
    folder = tmp / "x"
    folder.mkdir()
    (folder / name).write_bytes(data)
    sess._upload_folder = str(folder)
    sess.state.oa_messages.append(("user", f"📄 Uploaded `{name}`"))
    return _propose_or_defer(sess, folder, propose)