# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiKeyBody": {
        "properties": {
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider"
        ],
        "title": "ApiKeyBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Message": {
        "properties": {
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "text"
        ],
        "title": "Message",
        "type": "object"
      },
      "ModeBody": {
        "properties": {
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "mode"
        ],
        "title": "ModeBody",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP access to the Digicities platform backend: the same functions the Streamlit app calls in-process, exposed for the React frontend (`digicities-frontend`) and any other client.\n\nAlmost every route is scoped to a workspace via `/api/workspaces/{workspace_id}`; start with **workspaces** to find or create one.\n\nAuthentication is off by default (open local mode). With `API_AUTH_ENABLED` set, every route requires an `Authorization: Bearer <token>` header (Keycloak token; see `apps/api/auth.py`).",
    "title": "Digicities API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workspaces/{workspace_id}/agent/api-key": {
      "post": {
        "description": "Set (and persist) an LLM provider key from the settings panel; returns the\nupdated settings (key status never echoes the key itself).",
        "operationId": "set_api_key_api_workspaces__workspace_id__agent_api_key_post",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Set Api Key Api Workspaces  Workspace Id  Agent Api Key Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Api Key",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/agent/chats": {
      "get": {
        "description": "Persisted conversations for this workspace.",
        "operationId": "chats_api_workspaces__workspace_id__agent_chats_get",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Chats Api Workspaces  Workspace Id  Agent Chats Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chats",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/agent/commands": {
      "get": {
        "description": "The commands that can run in the conversation's current step, for the\nchat's command list: ``{\"stage\", \"commands\": [{key, area, form, usage,\ndescription, when_to_use, example, example_result, states, slots: [{name,\nkind, label, help, choices: [{value, label}]}]}]}``. Each command says when to\nuse it and what its example does; each slot says what goes in it. Slot\nchoices come from the workspace (classes, attributes, scenarios, instances,\nthe numbered decisions) and the ontology (the link predicates under\n``linksComponent``). Generated from the agent's command registry, the same\nlist its router and help use.",
        "operationId": "commands_api_workspaces__workspace_id__agent_commands_get",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Commands Api Workspaces  Workspace Id  Agent Commands Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commands",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/agent/message": {
      "post": {
        "operationId": "message_api_workspaces__workspace_id__agent_message_post",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Message"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Message Api Workspaces  Workspace Id  Agent Message Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Message",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/agent/message/stream": {
      "get": {
        "description": "Server-sent events: `token` as the LLM writes, then `result` with the full turn.\n\nGET exists for EventSource clients; long messages should use the POST\nvariant so the text rides in the body, not the query string / access logs.",
        "operationId": "message_stream_api_workspaces__workspace_id__agent_message_stream_get",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "text",
            "required": true,
            "schema": {
              "title": "Text",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Message Stream",
        "tags": [
          "agent"
        ]
      },
      "post": {
        "description": "Same SSE stream, message in the request body (fetch + ReadableStream).",
        "operationId": "message_stream_post_api_workspaces__workspace_id__agent_message_stream_post",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Message"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Message Stream Post",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/agent/mode": {
      "post": {
        "operationId": "set_mode_api_workspaces__workspace_id__agent_mode_post",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Set Mode Api Workspaces  Workspace Id  Agent Mode Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Mode",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/agent/model": {
      "post": {
        "operationId": "set_model_api_workspaces__workspace_id__agent_model_post",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Set Model Api Workspaces  Workspace Id  Agent Model Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Model",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/api/workspaces/{workspace_id}/agent/propose/stream": {
      "get": {
        "description": "Server-sent events: map the folder recorded by ``upload(propose=false)``,\nthen ``result`` with the full state and ``done``. 409 when there is nothing\nwaiting to be mapped.",
        "operationId": "propose_stream_api_workspaces__workspace_id__agent_propose_stream_get",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Propose Stream",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/api/workspaces/{workspace_id}/agent/settings": {
      "get": {
        "description": "Settings-panel data: per-provider key status + the read-only folder where the\nagent stores this workspace's artifacts (chats, ingestion report, provenance).",
        "operationId": "settings_api_workspaces__workspace_id__agent_settings_get",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Settings Api Workspaces  Workspace Id  Agent Settings Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Settings",
        "tags": [
          "agent"
        ]
      }
    },
...
  },
...
}